requirements.txt
requests==2.25.1
aiohttp==3.7.4.post0
pandas==1.2.4
numpy==1.20.3
openai==0.8.5
langchain
crewai
python-dotenv
Flask==2.0.1
Flask-Cors==3.0.10
